move op_params to common
