+++ selfdrive/updated.py
@@ -9,7 +9,7 @@ import os
 
 from common.params import Params
 from selfdrive.swaglog import cloudlog
-from selfdrive.op_params import opParams
+from common.op_params import opParams
 import selfdrive.messaging as messaging
 from selfdrive.services import service_list
